Extract PIC parsing into new copybook-pic microcrate
