+++ copybook-core/Cargo.toml
@@ -27,6 +27,7 @@ copybook-error-reporter.workspace = true
 copybook-overflow.workspace = true
 copybook-utils.workspace = true
 copybook-dialect.workspace = true
+copybook-pic.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 serde_plain.workspace = true
